[#69-Expand story initialization abilities] Adds `IntoStory` trait
This trait can be used in places where we don't require the exact instance of `Story`, just a type passthrough.

This type gives us the ability to run different initializers for the story.

We remove `AnyStory` as it's not required for anything and was just leftover
after failed experiment really.